[BPF] Support to compile once and run everywhere
This is a basic implementation for compiler support for
compile once and run everywhere. A lot of things are missing:
  . multi-dimensional array support
  . more testing needed
But it shows what the interface will look like.

  -bash-4.4$ cat trace.c
  struct pt_regs {
    long arg1;
  };

  struct sk_buff {
    int i;
    struct net_device *dev;
  };

  #define _(P) ({typeof(P) val = 0; bpf_probe_read(&val, sizeof(val), &P); val;})

  static int (*bpf_probe_read)(void *dst, int size, void *unsafe_ptr) =
          (void *) 4;

  int bpf_prog(struct pt_regs *ctx) {
    struct net_device *dev = 0;

    // ctx->arg1 does not need bpf_probe_read
    bpf_probe_read(&dev, sizeof(dev), &((struct sk_buff *)ctx->arg1)->dev);
    return dev != 0;
  }

In the above, we want to translate the third argument of
bpf_probe_read() as relocations.

  -bash-4.4$ clang -target bpf -O2 -g -S trace.c

The compiler will generate another subsession in .BTF.ext.
The tentative name is called OffsetReloc.
   BPFOffsetReloc Size
   struct SecLOffsetReloc for ELF section #1
   A number of struct BPFOffsetReloc for ELF section #1
   struct SecOffsetReloc for ELF section #2
   A number of struct BPFOffsetReloc for ELF section #2
   ...

  struct BPFOffsetReloc {
    uint32_t InsnOffset;    ///< Byte offset in this section
    uint32_t TypeID;        ///< TypeID for the relocation
    uint32_t OffsetNameOff; ///< The string to traverse types
  };

For the above test case, one record will be generated:
        .long   .Ltmp3    // insn
        .long   10        // type id
        .long   184       // type decode string: "0:1"

In string table:
        .ascii  "0:1"                  # string offset=184
The offset can be calculated as:
  size of 0 types + the 2nd member offset (0 representing the 1st member) of the structure.

The code:
  .Ltmp3:
        r2 = 8
        .loc    1 19 62                 # trace.c:19:62
  .Ltmp4:
        r3 = *(u64 *)(r1 + 0)
        r3 += r2
        r1 = r10
  .Ltmp5:
  .Ltmp6:
        .loc    1 16 30                 # trace.c:16:30
  .Ltmp7:
        r1 += -8
        .loc    1 19 3                  # trace.c:19:3
  .Ltmp8:
        r2 = 8
        call 4

The "r2 = 8" is the instruction for relocation. The compiler
has computed offset "8" based on the structures it can see.
Loader needs to adjust it if it changes on the host.

Compiling to generate object code and disassemble:
    0000000000000000 bpf_prog:
       0:       b7 02 00 00 00 00 00 00         r2 = 0
       1:       7b 2a f8 ff 00 00 00 00         *(u64 *)(r10 - 8) = r2
       2:       b7 02 00 00 08 00 00 00         r2 = 8
       3:       79 13 00 00 00 00 00 00         r3 = *(u64 *)(r1 + 0)
       4:       0f 23 00 00 00 00 00 00         r3 += r2
       5:       bf a1 00 00 00 00 00 00         r1 = r10
       6:       07 01 00 00 f8 ff ff ff         r1 += -8
       7:       b7 02 00 00 08 00 00 00         r2 = 8
       8:       85 00 00 00 04 00 00 00         call 4
Instruction #2 is the one loader needs to adjust if types do not match.

Currently, if type string like "0:<name>:0:0" where everything is "0" except
the type name, the llvm/clang will omit the getmemberptr IR and directly using
the base pointer since the offset will be 0 any way. This won't work
perfectly for us since we lose type information. (Theoretically BTF
could trace down the type, always tracing the first element of
struct/union or the first element of array, stops when the type size
is equal to the second argument of bpf_probe_read, but this rather
looks like hacky.

I implemented a hook in BPF backend to disable this optimization
so if users compile with "clang -target bpf", the getelementptr
will be always preserved. This should be okay since once we have
vmlinux.h, we should just encourage user to use "clang -target bpf"
directly since no kernel headers are involved.

WARNING:
union support is incomplete. If a member access starts with an
union access, it will not be transformed as the llvm IR directly
went to the members of union.

Signed-off-by: Yonghong Song <yhs@fb.com>