[BPF] Support to compile once and run everywhere
This is a basic implementation for compiler support for
compile once and run everywhere. Note that this patch
has not been extensively tested.

But it shows what the interface will look like.

struct pt_regs {
  long arg1;
  long arg2;
};

struct sk_buff {
  int i;
  struct net_device *dev;
};

static int (*bpf_probe_read)(void *dst, int size, void *unsafe_ptr) =
        (void *) 4;

extern __attribute__((section(".BPF.patchable_externs"))) unsigned __kernel_version;

int bpf_prog(struct pt_regs *ctx) {
  struct net_device *dev = 0;

  // ctx->arg* does not need bpf_probe_read
  if (__kernel_version >= 41608)
    bpf_probe_read(&dev, sizeof(dev), &((struct sk_buff *)ctx->arg1)->dev);
  else
    bpf_probe_read(&dev, sizeof(dev), &((struct sk_buff *)ctx->arg2)->dev);
  return dev != 0;
}

In the above, we want to translate the third argument of
bpf_probe_read() as relocations.

  -bash-4.4$ clang -target bpf -O2 -g -S trace.c

The compiler will generate two new subsections in .BTF.ext.
The tentative name is called OffsetReloc and ExternReloc.
OffsetReloc is to record the structure member offset operations,
and ExternalReloc is to record the external globals where
only u8, u16, u32 and u64 are supported.

   BPFOffsetReloc Size
   struct SecLOffsetReloc for ELF section #1
   A number of struct BPFOffsetReloc for ELF section #1
   struct SecOffsetReloc for ELF section #2
   A number of struct BPFOffsetReloc for ELF section #2
   ...
   BPFExternReloc Size
   struct SecExternReloc for ELF section #1
   A number of struct BPFExternReloc for ELF section #1
   struct SecExternReloc for ELF section #2
   A number of struct BPFExternReloc for ELF section #2

  struct BPFOffsetReloc {
    uint32_t InsnOffset;    ///< Byte offset in this section
    uint32_t TypeID;        ///< TypeID for the relocation
    uint32_t OffsetNameOff; ///< The string to traverse types
    uint32_t Dependency;    ///< Depending on another BPFOffsetReloc
  };

  struct BPFExternReloc {
    uint32_t InsnOffset;    ///< Byte offset in this section
    uint32_t ExternNameOff; ///< The string for external variable
  };

Note that only externs with attribute section ".BPF.patchable_externs"
are considered for Extern Reloc which will be patched by bpf loader
right before the load.

For the above test case, two offset records and one extern record
will be generated:
  OffsetReloc records:
        .long   .Ltmp12                 # Insn Offset
        .long   7                       # TypeId
        .long   239                     # Type Decode String
        .long   0                       # Dependency
        .long   .Ltmp18                 # Insn Offset
        .long   7                       # TypeId
        .long   239                     # Type Decode String
        .long   0                       # Dependency

  ExternReloc record:
        .long   .Ltmp5                  # Insn Offset
        .long   164                     # External Variable

In string table:
        .ascii  "0:0"                   # string offset=239
        .ascii  "__kernel_version"      # string offset=164

The offset can be calculated as:
  size of 0 types + the 2nd member offset (0 representing the 1st member) of the structure.

The code:
.Ltmp5:
.Ltmp6:
        r2 = 0
        r3 = 41608
.Ltmp7:
.Ltmp8:
        .loc    1 22 7 is_stmt 0        # trace.c:22:7
.Ltmp9:
        if r3 > r2 goto LBB0_2
.Ltmp10:
.Ltmp11:
        .loc    1 0 7                   # trace.c:0:7
.Ltmp12:
        r2 = 8
.Ltmp13:
        .loc    1 23 64 is_stmt 1       # trace.c:23:64
.Ltmp14:
.Ltmp15:
        r3 = *(u64 *)(r1 + 0)
        goto LBB0_3
.Ltmp16:
.Ltmp17:
LBB0_2:
        .loc    1 0 64 is_stmt 0        # trace.c:0:64
.Ltmp18:
        r2 = 8
        .loc    1 25 64 is_stmt 1       # trace.c:25:64
.Ltmp19:
        r3 = *(u64 *)(r1 + 8)
.Ltmp20:
.Ltmp21:
LBB0_3:
        .loc    1 0 64 is_stmt 0        # trace.c:0:64
        r3 += r2
        r1 = r10
.Ltmp22:
.Ltmp23:
.Ltmp24:
        r1 += -8
        r2 = 8
        call 4

For instruction .Ltmp12 and .Ltmp18, "r2 = 8", the number
8 is the structure offset based on the current BTF.
Loader needs to adjust it if it changes on the host.

For instruction .Ltmp5, "r2 = 0", the external variable
got a default value 0, loader needs to supply an appropriate
value for the particular host.

Compiling to generate object code and disassemble:
    0000000000000000 bpf_prog:
       0:       b7 02 00 00 00 00 00 00         r2 = 0
       1:       7b 2a f8 ff 00 00 00 00         *(u64 *)(r10 - 8) = r2
       2:       b7 02 00 00 00 00 00 00         r2 = 0
       3:       b7 03 00 00 88 a2 00 00         r3 = 41608
       4:       2d 23 03 00 00 00 00 00         if r3 > r2 goto +3 <LBB0_2>
       5:       b7 02 00 00 08 00 00 00         r2 = 8
       6:       79 13 00 00 00 00 00 00         r3 = *(u64 *)(r1 + 0)
       7:       05 00 02 00 00 00 00 00         goto +2 <LBB0_3>

0000000000000040 LBB0_2:
       8:       b7 02 00 00 08 00 00 00         r2 = 8
       9:       79 13 08 00 00 00 00 00         r3 = *(u64 *)(r1 + 8)

0000000000000050 LBB0_3:
      10:       0f 23 00 00 00 00 00 00         r3 += r2
      11:       bf a1 00 00 00 00 00 00         r1 = r10
      12:       07 01 00 00 f8 ff ff ff         r1 += -8
      13:       b7 02 00 00 08 00 00 00         r2 = 8
      14:       85 00 00 00 04 00 00 00         call 4

Instructions #2, #5 and #8 need relocation resoutions from the loader.

There is one change outside of the Target/BPF.
Currently, if type string like "0:0:0" where everything is "0"
the llvm/clang will omit the getmemberptr IR and directly using
the base pointer since the offset will be 0 any way. This won't work
perfectly for us since we lose type information. (Theoretically BTF
could trace down the type, always tracing the first element of
struct/union or the first element of array, stops when the type size
is equal to the second argument of bpf_probe_read, but this rather
looks like hacky.
I implemented a hook in BPF backend to disable this optimization
so if users compile with "clang -target bpf", the getelementptr
will be always preserved. This should be okay since once we have
vmlinux.h, we should just encourage user to use "clang -target bpf"
directly since no kernel headers are involved.

WARNING:
union support is incomplete. If a member access starts with an
union access, it will not be transformed as the llvm IR directly
went to the members of union.

Signed-off-by: Yonghong Song <yhs@fb.com>